Simplify
The goal of this patch is to provide a clearer structure to this Docker
project, as well as providing an easy way for people to update the
docker image.

The ``run.sh`` script was replaced with docker-compose which makes it
easier to tweak the runtime options. Adding a `slides.md` file in the
slide directory is now the only thing required to visualize your
presentation.

A comprehensive markdown presentation example was added in the `slides`
directory. It reproduces the upstream demo.html presentation.